+++ src/arta/models.py
@@ -1,37 +1,41 @@
-"""Pydantic model implementations.
+"""Pydantic model implementations."""
 
-Note: Having no "from __future__ import annotations" here is wanted (pydantic compatibility).
-"""
+from __future__ import annotations
 
-from typing import Annotated, Any, Callable, Dict, List, Optional
+from typing import Annotated, Any, Callable, Optional
+from warnings import warn
 
 import pydantic
 from pydantic.version import VERSION
 
 from arta.utils import ParsingErrorStrategy
 
-if not VERSION.startswith("1."):
+PYDANTIC_V1: bool = VERSION.startswith("1.")
+
+if not PYDANTIC_V1:
+    # Pydantic V2
+
     # ----------------------------------
     # For instantiation using rules_dict
     class RuleRaw(pydantic.BaseModel):
         """Pydantic model for validating a rule."""
 
-        condition: Optional[Callable]
-        condition_parameters: Optional[Dict[str, Any]]
+        condition: Optional[Callable] = None
+        condition_parameters: Optional[dict[str, Any]] = None
         action: Callable
-        action_parameters: Optional[Dict[str, Any]]
+        action_parameters: Optional[dict[str, Any]] = None
 
         model_config = pydantic.ConfigDict(extra="forbid")
 
-    class RulesGroup(pydantic.RootModel):  # noqa
+    class RulesGroup(pydantic.RootModel):
         """Pydantic model for validating a rules group."""
 
-        root: Dict[str, RuleRaw]
+        root: dict[str, RuleRaw]
 
-    class RulesDict(pydantic.RootModel):  # noqa
+    class RulesDict(pydantic.RootModel):
         """Pydantic model for validating rules dict instanciation."""
 
-        root: Dict[str, RulesGroup]
+        root: dict[str, RulesGroup]
 
     # ----------------------------------
     # For instantiation using config_path
@@ -40,14 +44,14 @@ if not VERSION.startswith("1."):
 
         description: str
         validation_function: str
-        condition_parameters: Optional[Dict[str, Any]] = None
+        condition_parameters: Optional[dict[str, Any]] = None
 
     class RulesConfig(pydantic.BaseModel):
         """Pydantic model for validating a rule group from config file."""
 
         condition: Optional[str] = None
         simple_condition: Optional[str] = None
-        action: Annotated[str, pydantic.StringConstraints(to_lower=True)]  # type: ignore
+        action: Annotated[str, pydantic.StringConstraints(to_lower=True)]
         action_parameters: Optional[Any] = None
 
         model_config = pydantic.ConfigDict(extra="allow")
@@ -55,37 +59,35 @@ if not VERSION.startswith("1."):
     class Configuration(pydantic.BaseModel):
         """Pydantic model for validating configuration files."""
 
-        conditions: Optional[Dict[str, Condition]] = None
-        conditions_source_modules: Optional[List[str]] = None
-        actions_source_modules: List[str]
-        custom_classes_source_modules: Optional[List[str]] = None
-        condition_factory_mapping: Optional[Dict[str, str]] = None
-        rules: Dict[str, Dict[str, Dict[str, RulesConfig]]]
+        conditions: Optional[dict[str, Condition]] = None
+        conditions_source_modules: Optional[list[str]] = None
+        actions_source_modules: list[str]
+        custom_classes_source_modules: Optional[list[str]] = None
+        condition_factory_mapping: Optional[dict[str, str]] = None
+        rules: dict[str, dict[str, dict[Annotated[str, pydantic.StringConstraints(to_upper=True)], RulesConfig]]]
         parsing_error_strategy: Optional[ParsingErrorStrategy] = None
 
-        model_config = pydantic.ConfigDict(extra="ignore")
-
-        @pydantic.field_validator("rules", mode="before")  # noqa
-        def upper_key(cls, vl):  # noqa
-            """Validate and uppercase keys for RulesConfig"""
-            for k, v in vl.items():
-                for kk, vv in v.items():
-                    for key, rules in [*vv.items()]:
-                        if key != str(key).upper():
-                            del vl[k][kk][key]
-                            vl[k][kk][str(key).upper()] = rules
-            return vl
-
 else:
+    # Pydantic V1
+
+    warn(
+        (
+            "Soon, Pydantic V1 will no longer be compatible with Arta. "
+            "Please, migrate to Pydantic V2 (https://docs.pydantic.dev/latest/migration/)."
+        ),
+        DeprecationWarning,
+        stacklevel=2,
+    )
 
     class BaseModelV2(pydantic.BaseModel):
         """Wrapper to expose missed methods used elsewhere in the code"""
 
-        model_dump: Callable = pydantic.BaseModel.dict  # noqa
+        model_dump: Callable = pydantic.BaseModel.dict
 
         @classmethod
-        def model_validate(cls, obj):  # noqa
-            return cls.parse_obj(obj)  # noqa
+        def model_validate(cls, obj):
+            """Method mapping between V1 to V2."""
+            return cls.parse_obj(obj)
 
     # ----------------------------------
     # For instantiation using rules_dict
@@ -93,22 +95,22 @@ else:
         """Pydantic model for validating a rule."""
 
         condition: Optional[Callable]
-        condition_parameters: Optional[Dict[str, Any]]
+        condition_parameters: Optional[dict[str, Any]]
         action: Callable
-        action_parameters: Optional[Dict[str, Any]]
+        action_parameters: Optional[dict[str, Any]]
 
         class Config:
             extra = "forbid"
 
-    class RulesGroup(pydantic.BaseModel):  # type: ignore[no-redef] # noqa
+    class RulesGroup(pydantic.BaseModel):  # type: ignore[no-redef]
         """Pydantic model for validating a rules group."""
 
-        __root__: Dict[str, RuleRaw]  # noqa
+        __root__: dict[str, RuleRaw]
 
-    class RulesDict(BaseModelV2):  # type: ignore[no-redef]  # noqa
+    class RulesDict(BaseModelV2):  # type: ignore[no-redef]
         """Pydantic model for validating rules dict instanciation."""
 
-        __root__: Dict[str, RulesGroup]  # noqa
+        __root__: dict[str, RulesGroup]
 
     # ----------------------------------
     # For instantiation using config_path
@@ -117,7 +119,7 @@ else:
 
         description: str
         validation_function: str
-        condition_parameters: Optional[Dict[str, Any]]
+        condition_parameters: Optional[dict[str, Any]]
 
     class RulesConfig(BaseModelV2):  # type: ignore[no-redef]
         """Pydantic model for validating a rule group from config file."""
@@ -133,13 +135,10 @@ else:
     class Configuration(BaseModelV2):  # type: ignore[no-redef]
         """Pydantic model for validating configuration files."""
 
-        conditions: Optional[Dict[str, Condition]]
-        conditions_source_modules: Optional[List[str]]
-        actions_source_modules: List[str]
-        custom_classes_source_modules: Optional[List[str]]
-        condition_factory_mapping: Optional[Dict[str, str]]
-        rules: Dict[str, Dict[str, Dict[pydantic.constr(to_upper=True), RulesConfig]]]  # type: ignore
+        conditions: Optional[dict[str, Condition]]
+        conditions_source_modules: Optional[list[str]]
+        actions_source_modules: list[str]
+        custom_classes_source_modules: Optional[list[str]]
+        condition_factory_mapping: Optional[dict[str, str]]
+        rules: dict[str, dict[str, dict[pydantic.constr(to_upper=True), RulesConfig]]]  # type: ignore
         parsing_error_strategy: Optional[ParsingErrorStrategy]
-
-        class Config:
-            extra = "ignore"